requirements.txt
setuptools>=69.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
mlflow==2.16.2
matplotlib==3.8.4
seaborn==0.13.2
jupyter==1.0.0
joblib==1.4.2
fastapi==0.111.0
uvicorn==0.30.1
azure-storage-blob==12.20.0
python-dotenv==1.0.1
